JFR test fixes
